Create README.md
What is the effect of the quality of bureaucracy in promoting innovation in Brazilian municipalities? This paper responds to this question by proposing a new Municipal Bureaucracy Quality Index (IQBM), measured from three dimensions: level of schooling, outsourcing and politicization. The central hypothesis of the thesis holds that the more qualified the bureaucracy is the greater the probability of a Brazilian municipality to innovate. In addition, other variables have been identified that contribute to increase the chance of promoting a new idea, such as political alignment, municipal revenue, and some characteristics of mayors such as age, schooling and the possibility of re-election. Methodologically, the research design uses the technique of principal components analysis to create the IQBM and the logistic regression to identify the influence exerted by the indicator on the chances of the municipalities to innovate. An original database was built, compiled from information provided by the Municipal Basic Information Survey (MUNIC) conducted by IBGE, Superior Electoral Tribunal, National Treasury and SEBRAE Nacional. The main results indicate that: (1) municipalities with more skilled bureaucracy have a greater chance of promoting innovative policies or programs, and (2) that there is a positive influence of political alignment and municipal revenue on the chances of these subnational units innovating. This research contributes to the literature on the subject because it presents an original character when creating an indicator that measures the quality of bureaucracy for the Brazilian municipalities (IQBM) inferring about the probability of them innovating.

Keywords: Innovation in the Public Sector. Public Bureaucracy. Public Policies